+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Holidays are now shown in the Weekly view day title bars.
 - The holiday text shown in the title bars now uses a new custom CSS block called "light-title-bar-text" (which is lighter).
 - Added new setting "showHolidays", which states if the holidays should be shown (defaults to true).
+- Added new option "Show holidays in the main display and title bars" to Configuration > Display.
 - 
 
 ### Version 1.3.9: